requirement.txt
torch
torchtext
torch-ignite
pandas
numpy
